% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeRsphere.R
\name{normalizeRsphere}
\alias{normalizeRsphere}
\title{Normalize values on a sphere}
\usage{
normalizeRsphere(x, r = 1)
}
\arguments{
\item{x}{A vector or a matrix with values to be normalized.}

\item{r}{The diameter of a sphere, default 1.}
}
\value{
It returns a data frame with normalized values.
}
\description{
Normalizes values of a vector such that the sum of squared elements equal to \eqn{r^2}.
}
\examples{
normalizeRsphere(x = c(1, 0.5, 0.4))
}
\author{
Marjan Cugmas
}
