% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popularity.R
\name{popularity}
\alias{popularity}
\title{Popularity mechanism}
\usage{
popularity(X, actor = NULL)
}
\arguments{
\item{X}{Binary network; of class \code{matrix}.}

\item{actor}{Not used by the function, set to \code{NULL}. Necessary for using within other functions, e.g. nemBM.}
}
\value{
A vector with the normalized popularity mechanism, cacluated for each unit.
}
\description{
Calculate the normalized network statistic according to the popularity mechanism.
}
\examples{
X <- matrix(sample(c(0,1), size = 9**2, replace = TRUE), nrow = 9)
popularity(X)
}
\references{
Cugmas, M., & Žiberna, A. (2022). Approaches to blockmodeling dynamic networks: a Monte Carlo simulation study. Social Networks, in print.
}
\author{
Marjan Cugmas and Aleš Žiberna
}
