% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomizePartition.R
\name{randomizePartition}
\alias{randomizePartition}
\title{Randomize a partition}
\usage{
randomizePartition(partition, p, checkSelected = FALSE)
}
\arguments{
\item{partition}{Initial partition in a vector format. Each unique value (positive integers) represents one cluster.}

\item{p}{The share of relocated units.}

\item{checkSelected}{If \code{TRUE} (default is \code{FALSE}) a given unit can be relocated only once.}
}
\value{
A partition (in a vector format).
}
\description{
It randomizes a partition by randomly relocating a given share of units between the clusters. The group sizes are preserved.
}
\examples{
randomizePartition(partition = c(1, 1, 1, 1, 2, 2, 2, 2, 3, 3, 3, 3), p = 0.3)
}
\author{
Marjan Cugmas and Aleš Žiberna
}
